Various minor fixes

+++ SelfControl Killer/AppDelegate.m
@@ -24,7 +24,7 @@
 - (IBAction)killButtonClicked:(id)sender {
 	AuthorizationRef authorizationRef;
 	char* helperToolPath = [self selfControlKillerHelperToolPathUTF8String];
-	int helperToolPathSize = strlen(helperToolPath);
+	unsigned long helperToolPathSize = strlen(helperToolPath);
 	AuthorizationItem right = {
 		kAuthorizationRightExecute,
 		helperToolPathSize,
